+++ src/app/math/index.ts
@@ -1,4 +1,4 @@
-import { MathUtilMathJs } from './impl/mathUtilMathJs';
+import { MathUtilComputeIo } from './impl/mathUtilComputeIo';
 import { MathFunctions } from './util/math.function';
 
 /**
@@ -7,11 +7,33 @@ import { MathFunctions } from './util/math.function';
 
 /**
  *
- * Returns the smallest number in the array
+ * Returns the smallest number from numbers
  *
- * @param {!number[]} numbers - The array of numbers
+ * @param {!number[]} numbers - The numbers
+ * @returns {number} The smallest number
+ */
+export function max(...numbers: number[]): number {
+  return new MathFunctions(new MathUtilComputeIo()).max(numbers);
+}
+
+/**
+ *
+ * Returns the smallest number from numbers
+ *
+ * @param {!number[]} numbers - The numbers
  * @returns {number} The smallest number
  */
 export function min(...numbers: number[]): number {
-  return new MathFunctions(new MathUtilMathJs()).min(numbers);
+  return new MathFunctions(new MathUtilComputeIo()).min(numbers);
+}
+
+/**
+ *
+ * Returns the sum from numbers
+ *
+ * @param {!number[]} numbers - The numbers
+ * @returns {number} The sum
+ */
+export function sum(...numbers: number[]): number {
+  return new MathFunctions(new MathUtilComputeIo()).sum(numbers);
 }