+++ src/app/math/util/math.function.ts
@@ -4,8 +4,16 @@ export class MathFunctions {
 
   constructor(private mathUtilImpl: MathUtil) { }
 
+  max(numbers: number[]): number {
+    return this.mathUtilImpl.max(numbers);
+  }
+
   min(numbers: number[]): number {
     return this.mathUtilImpl.min(numbers);
+  }
+
+  sum(numbers: number[]): number {
+    return this.mathUtilImpl.sum(numbers);
   }
 
 }