+++ src/app/math/util/math.util.interface.ts
@@ -1,3 +1,5 @@
 export interface MathUtil {
+  max(numbers: number[]): number;
   min(numbers: number[]): number;
+  sum(numbers: number[]): number;
 }