+++ src/test/integration/math/mathjs/index.spec.ts
@@ -1,6 +1,0 @@
-import { MathUtilMathJs } from '../../../../app/math/impl/mathUtilMathJs';
-import { MinTest } from '../min';
-
-const mathUtilMathJs = new MathUtilMathJs();
-
-new MinTest(mathUtilMathJs).runTest();